part4 done!

+++ part4/blog/controllers/blogs.js
@@ -1,32 +1,57 @@
 const blogRouter = require('express').Router()
-const { request, response } = require('../app')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const { userExtractor } = require('../utils/middleware')
 
 // get all blogs
 blogRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({})
+    .populate('user',
+    { username: 1, name: 1 })
+
     response.json(blogs)
   })
 
 // add a blog
-blogRouter.post('/', async (request, response) => {
-    const blog = request.body
+blogRouter.post('/', userExtractor, async (request, response) => {
+  const blog = request.body
+  const user = request.user
 
-    const blogToPost = new Blog({
-      title: blog.title,
-      author: blog.author,
-      url: blog.url,
-      likes: blog.likes
-    })
-  
-    const blogSaved = await blogToPost.save()
-    response.status(201).json(blogSaved)
+  const blogToPost = new Blog({
+    title: blog.title,
+    author: blog.author,
+    url: blog.url,
+    likes: blog.likes,
+    user: user._id
+  })
+
+  const blogSaved = await blogToPost.save()
+  // from promise, take post id, add to users blogs array field
+  user.blogs = user.blogs.concat(blogSaved._id)
+  // save via promise
+  await user.save()
+  response.status(201).json(blogSaved)
   })
 
 // delete blog
-blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogRouter.delete('/:id', userExtractor, async (request, response) => {
+
+  // get user id
+  const token = request.token
+  
+  // get user for blog 
+  const user = request.user
+
+  if (user.id === token.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    return response.status(401).json({
+      error: 'token missing or invalid'
+    })
+  }
 })
 
 // edit a blog -> likes can change
